# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESERVA DE AULAS PARA EXAMENES ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/notificacion": {
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Creacion de una nueva notificacion de un docente",
                "operationId": "64ad5b0c357db5a04ed2c4aa596c3141",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mensajeNotificacion": {
                                        "type": "string"
                                    },
                                    "id_docente": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/notificacion/{idNotificacion}": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Obtencion de una notificacion apartir de un id",
                "operationId": "939c6d636a0c4757f7c10efec66e4ab6",
                "parameters": [
                    {
                        "name": "idNotificacion",
                        "in": "path",
                        "description": "Id de la Notificacion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}